Move the remaining clipboard-related tests under System.Windows.Form.tests to the new class
